Minor additions
added an analogue of the global Pipeline helper function ‘use’, but suitable for use with non-objects (value types), might be usefull when trycheck something inside a struct and throw in case of failure, or just let the whole struct to pass through in case of success. Also made the ‘use’ itself usable with value types.